requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
bitarray==2.9.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
Cython==3.0.8
decorator==4.4.2
dlib==19.24.2
-e git+https://github.com/pytorch/fairseq@fad2c4d1ebe14d974876de52dcb06db6d99b0b4a#egg=fairseq
ffmpeg-python==0.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
google-auth==2.26.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
hydra-core==1.3.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.3
lightning-utilities==0.10.1
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
portalocker==2.8.2
proglog==0.1.10
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pyDeprecate==0.3.1
pydub==0.25.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytorch-lightning==1.5.0
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
soundfile==0.12.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.7.10
torch==2.1.2+cu118
torchaudio==2.1.2
torchmetrics==1.3.0.post0
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
